requirements_test.txt
# Core testing dependencies
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
hypothesis>=6.0.0,<7.0.0
build>=1.0.0,<2.0.0
twine>=4.0.0,<5.0.0

# Core dependencies
numpy>=1.20.0
pandas>=2.0.0
scikit-learn>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0

# For visualization tests
matplotlib>=3.5.0
plotly>=5.10.0

# Optional dependencies for functional tests
# Install these when running full test suite
umap-learn>=0.5.3,<0.6.0
sentence-transformers>=2.2.0,<3.0.0
hdbscan>=0.8.29,<0.9.0
gensim>=4.1.0,<5.0.0